Update code for omd 2.0.0~alpha1
The interface of omd has already changed since 2.0.0~alpha1,
this is based on @jfrolich PR https://github.com/ocaml/omd/pull/215
and is intended as an intermediate step to ease the port to the next
release of omd. Issue #5 will not be closed for the time being.

In particular there is a regression in the way we deal with blockquotes.
These need to be addressed anyway, see #2.

Signed-off-by: Marcello Seri <marcello.seri@gmail.com>